Finished remaining chart views. Removed old demo.
