+++ demo/src/views/bar-view.js
@@ -1,6 +1,6 @@
 import React from "react";
-import { View, ScrollView } from "react-native";
-// import { VictoryChart } from "victory-native";
+import { ScrollView } from "react-native";
+import { VictoryBar, VictoryGroup, VictoryStack } from "victory-native";
 import viewStyles from "../styles/view-styles";
 
 export default class extends React.Component {
@@ -11,7 +11,51 @@ export default class extends React.Component {
   render() {
     return (
       <ScrollView style={viewStyles.container}>
-        <View />
+        <VictoryBar />
+
+        <VictoryBar
+          data={[
+            { x: 1, y: 1 },
+            { x: 2, y: 2 },
+            { x: 3, y: 3 },
+            { x: 4, y: 2 },
+            { x: 5, y: 1 }
+          ]}
+        />
+
+        <VictoryGroup
+          width={300}
+          height={375}
+          offset={20}
+          colorScale={"qualitative"}
+        >
+          <VictoryBar data={[{ x: 1, y: 1 }, { x: 2, y: 2 }, { x: 3, y: 3 }]} />
+          <VictoryBar data={[{ x: 1, y: 2 }, { x: 2, y: 1 }, { x: 3, y: 1 }]} />
+          <VictoryBar data={[{ x: 1, y: 3 }, { x: 2, y: 4 }, { x: 3, y: 2 }]} />
+        </VictoryGroup>
+
+        <VictoryStack width={300} height={375} colorScale={"qualitative"}>
+          <VictoryBar data={[{ x: 1, y: 1 }, { x: 2, y: 2 }, { x: 3, y: 3 }]} />
+          <VictoryBar data={[{ x: 1, y: 2 }, { x: 2, y: 1 }, { x: 3, y: 1 }]} />
+          <VictoryBar data={[{ x: 1, y: 3 }, { x: 2, y: 4 }, { x: 3, y: 2 }]} />
+        </VictoryStack>
+
+        <VictoryBar
+          height={375}
+          padding={75}
+          style={{
+            data: {
+              fill: data => (data.y > 2 ? "red" : "blue")
+            }
+          }}
+          data={[
+            { x: 1, y: 1 },
+            { x: 2, y: 2 },
+            { x: 3, y: 3 },
+            { x: 4, y: 2 },
+            { x: 5, y: 1 }
+          ]}
+        />
       </ScrollView>
     );
   }
